refactor: Update environment imports and restore environment files for production and development
